lkl: fix conflicted file name on case-insensitive fs build
macOS uses case-insensitive fs (apfs/hfs) by default and travis also use
it.  When we added musl's _exit.c as a build file, another file, _Exit.c
conflicts with the new file on case-insensitive file system.  This
happens when we created our own libc.a, extracting musl's files in a
single directory.

This commit fixes this issue by renaming a file as a fixup.

Fixes: 4ecff1b8 ("lkl: fix exit syscall usage after main function")

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>